modify source

+++ Source/Ack/VPSocketAck.h
@@ -1,9 +1,9 @@
 //
 //  VPSocketAck.h
-//  VPSocketIO
+//  IFMSocketIO
 //
-//  Created by Vasily Popov on 9/26/17.
-//  Copyright © 2017 Vasily Popov. All rights reserved.
+//  Created by yangguang on 2018/7/24.
+//  Copyright © 2018年 bngj. All rights reserved.
 //
 
 #import <Foundation/Foundation.h>
@@ -11,10 +11,8 @@
 
 @interface VPSocketAck : NSObject
 
-@property (nonatomic, readonly) int ack;
+@property (nonatomic, readonly) NSInteger ack;
 @property (nonatomic, strong, readonly) VPScoketAckArrayCallback callback;
 
 -(instancetype)initWithAck:(int)ack andCallBack:(VPScoketAckArrayCallback)callback;
-
 @end
-
